Move claimtrie-related service/handlers to jsonrpc_claimtrie.go.
